homepage/requirements.txt
# =============================================================================
# HOMELAB HOMEPAGE - PYTHON DEPENDENCIES
# =============================================================================
# Production dependencies for the homepage API key manager and utilities

# Core dependencies
cryptography>=41.0.0
requests>=2.31.0
PyYAML>=6.0.1

# Security and validation
python-dotenv>=1.0.0
certifi>=2023.7.22

# Logging and monitoring
structlog>=23.1.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.0.0
# mypy>=1.5.0 